requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
requests
together
